+++ src/react/internal/cache/QueryReference.ts
@@ -158,6 +158,7 @@ export class InternalQueryReference<TData = unknown> {
   private reject: ((error: unknown) => void) | undefined;
 
   private references = 0;
+  private softReferences = 0;
 
   constructor(
     observable: ObservableQuery<TData, any>,
@@ -241,9 +242,28 @@ export class InternalQueryReference<TData = unknown> {
       disposed = true;
       this.references--;
 
-      // Wait before fully disposing in case the app is running in strict mode.
+      if (!this.references) {
+        this.dispose();
+      }
+    };
+  }
+
+  softRetain() {
+    this.softReferences++;
+    let disposed = false;
+
+    return () => {
+      // Tracking if this has already been called helps ensure that
+      // multiple calls to this function won't decrement the reference
+      // counter more than it should. Subsequent calls just result in a noop.
+      if (disposed) {
+        return;
+      }
+
+      disposed = true;
+      this.softReferences--;
       setTimeout(() => {
-        if (!this.references) {
+        if (!this.softReferences && !this.references) {
           this.dispose();
         }
       });
@@ -253,6 +273,7 @@ export class InternalQueryReference<TData = unknown> {
   didChangeOptions(watchQueryOptions: ObservedOptions) {
     return OBSERVED_CHANGED_OPTIONS.some(
       (option) =>
+        option in watchQueryOptions &&
         !equal(this.watchQueryOptions[option], watchQueryOptions[option])
     );
   }
@@ -377,7 +398,7 @@ export class InternalQueryReference<TData = unknown> {
     // to resolve the promise if `handleNext` hasn't been run to ensure the
     // promise is resolved correctly.
     returnedPromise
-      .then((result) => {
+      .then(() => {
         // In the case of `fetchMore`, this promise is resolved before a cache
         // result is emitted due to the fact that `fetchMore` sets a `no-cache`
         // fetch policy and runs `cache.batch` in its `.then` handler. Because
@@ -390,8 +411,16 @@ export class InternalQueryReference<TData = unknown> {
         // more information
         setTimeout(() => {
           if (this.promise.status === "pending") {
-            this.result = result;
-            this.resolve?.(result);
+            // Use the current result from the observable instead of the value
+            // resolved from the promise. This avoids issues in some cases where
+            // the raw resolved value should not be the emitted value, such as
+            // when a `fetchMore` call returns an empty array after it has
+            // reached the end of the list.
+            //
+            // See the following for more information:
+            // https://github.com/apollographql/apollo-client/issues/11642
+            this.result = this.observable.getCurrentResult();
+            this.resolve?.(this.result);
           }
         });
       })